+++ Assets/Scripts/Enemies and Projectiles/EnemyHugger.cs
@@ -117,7 +117,7 @@ public class EnemyHugger : Enemy
           return isAgro;
      }
 
-     public int currentHp()
+     public float currentHp()
      {
           return health.currentHealth;
      }
